Replaced QIconButton
QIconButton has moved to models.widgets._icon.
There were two classes called QIconButton and the public one made more sense.
Removed private QIconButton.
Replaced references to QIconButton with AIconButton.
