+++ src/amulet_editor/models/widgets/__init__.py
@@ -1,5 +1,5 @@
 from amulet_editor.models.widgets._card import QLinkCard, QPixCard
 from amulet_editor.models.widgets._draggable import QDragContainer, QDragIconButton
-from amulet_editor.models.widgets._icon import QIconButton, QSvgIcon, AStylableSvgWidget
+from amulet_editor.models.widgets._icon import QSvgIcon, QIconButton, AIconButton, AStylableSvgWidget
 from amulet_editor.models.widgets._label import QElidedLabel
 from amulet_editor.models.widgets._menu import QMenuWidget
