+++ src/amulet_editor/resources/themes/default.scss
@@ -551,23 +551,22 @@ QTextEdit[readOnly=true] {
     color: $on_surface;
 }
 
-QToolButton {
+QToolButton,
+QToolButton:pressed
+{
     background-color: $surface;
     color: $on_surface;
 }
-QToolButton:checked {
+QToolButton:checked,
+QIconButton:checked,
+AIconButton:checked
+{
     background-color: $secondary_x0_666;
     color: $on_secondary;
 }
-QToolButton:pressed {
-    background-color: $surface;
-    color: $on_surface;
-}
-QIconButton:checked {
-    background-color: $secondary_x0_666;
-    color: $on_secondary;
-}
-QIconButton:pressed {
+QIconButton:pressed,
+AIconButton:pressed
+{
     background-color: $primary;
     color: $on_primary;
 }
